blog/requirements.txt
asgiref==3.10.0
Django==3.2.16
pillow==12.0.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
